Updated jb_psa-populate_metadata_tables to execute the SQL scripts to first delete data from the ETL DB and to run the "adjustment" SQL scripts after the default data is insertd.
Changed data type of etl.column_meta.column_meta_id from smallint to
integer.

+++ database/dwh_etl_mgmt/dbms/postgresql/create-schema-dars-dwh_etl_mgmt-postgresql.sql
@@ -16,7 +16,7 @@ CREATE TABLE etl.cdc_timestamps (
 --
 
 CREATE TABLE etl.column_meta (
-    column_meta_id smallint NOT NULL,
+    column_meta_id integer NOT NULL,
     table_meta_id smallint NOT NULL,
     source_column_name character varying(80),
     target_column_name character varying(80),
